+++ CHANGES.rst
@@ -6,9 +6,12 @@ Changelog
 - new PATCH version for backwards compatible bug fixes
 
 
-0.4.9a0
+0.4.9
 -------
-2020-07-27: develop
+2020-07-27: feature release
+    - add command script
+    - add command after_success
+    - add command deploy
 
 
 0.4.8